Изменил модели данных для истории тестов

+++ app/src/main/java/com/example/mypsychologist/extensions/Extentions.kt
@@ -43,5 +43,5 @@ fun <A, B> Map<A, B>.containsKeys(vararg keys: A): Boolean = run {
     contains
 }
 
-fun Int.toPercent(maxValue: Int) =
+fun Float.toPercent(maxValue: Int) =
     this * 100 / maxValue
